twister: bugfix: Fix twister output
Fix twister output. "Test suite" used previously did not correspond
to what was counted (test scenarios). A test scenario is an entry from
testcase/sample.yaml under "tests:", which specify configuration used to
build a test. Changes in docs will follow later on. Also replace the
amount of preselectd platforms with the amount of platforms that
were not filtered out and actually tested.

Change-Id: I4d4553e14bf54279516982487b815e207f6b3674
Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>