pm: Avoid weak function clash
SoC hooks are defined as weak functions for the case where they are
not implemented by a SoC. The problem that may happen is when an
application define it as weak as well. In this case, no compilation
error is issued but the version that will be used depends on the build
order. Just remove the weak placeholder implementation.

Change-Id: Ifcd131c6d69ce94eb263882cf45227c55a5b8737
Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>