cmake: oneApi: add oneApi support on windows
The icx compiler of oneApi will invoke clang-cl on windows,
this is not supported in zephyr now, so change to use
clang directly.
And the clang from oneApi can't recognize those cross
compiling target variables of cmake, such as
CMAKE_C_COMPILER_TARGET, so used other cmake functions
to pass them to clang.

Change-Id: I6bd0cda326606de122e3c29d0eb6ab795fd71a73
Signed-off-by: Chen Peng1 <peng1.chen@intel.com>