boards: pinetime_devkit0: adjust spi flash compatible
We have a 'mxicy' entry in vendor-prefixes.txt for this vendor. Use it
to be consistent. Linux's
Documentation/devicetree/bindings/mtd/jedec,spi-nor.yaml file allows
both, but I'd like to see what happens if we try to be consistent in
Zephyr. There isn't a binding for this compatible in zephyr, so I'm
hopeful this won't break any upstream use cases.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>