MAINTAINERS: Add Bluetooth Audio section
Add a Bluetooth Audio section in the maintainers
file for the Bluetooth Audio (sub)subsystem.

The Bluetooth Audio stack is a significant size, and
there's a signficant amount of people working on it,
where many do not have proper review rights and
many that do not get automatically assigned
for reviews.

I've included Vudentz as he has been very active in
the early days.

I've excluded szymon-czapracki as he has not yet
contributed a significant amount, even if he is
involved.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>