wifi: fix build error on wifi_winc1500_nm_bsp_internal.h
Previous commit 579ca90e25271a693044667418acef69669f353e to
build wifi drivers as a library changed the include path for
the WINC1500 driver, which results in the include path being
local to the library. However, the Atmel HAL requires
wifi_winc1500_nm_bsp_internal.h to in the search path. So
change the include path to be global.

Fixes #43456

Signed-off-by: Daniel Leung <daniel.leung@intel.com>